Line ending fixes. Fixed bug in TServerList regarding logging. Slight codec improvement.

+++ gserver-v2/server/src/codec.cpp
@@ -42,7 +42,7 @@ int codec::limitfromtype(uint8_t type)
 		// If we found a valid type, set the limit.
 		if ( limits[i] == type )
 		{
-			m_limit = limits[i+1];
+			m_limit = (int32_t)limits[i+1];
 			return 0;
 		}
 	}
